Server/flask-kafka/agent2/compliance/requirements.txt
kafka-python
pymongo
python-dotenv
flask
requests
